refactor: cleanup stale files, move commands
Moving some of the code to more logical locations, cleaning up stale
files
