--- src/actions/actions.ts
@@ -1,8 +1,0 @@
-export enum Actions {
-  Unknown,
-  Init,
-  Check,
-  Config,
-  Chat,
-  Quit,
-}
