+++ src/commands/chat/chat.ts
@@ -1,7 +1,7 @@
-import { ExecutionContext } from "../lib/execution-context";
-import { Configuration } from "../configuration/configuration";
-import { executeChatPipeline } from "../chat-pipeline/chat-pipeline-completion-api";
-import { executeChatPipeline as executeAssistantPipeline } from "../chat-pipeline/chat-pipeline-assistant-api";
+import { ExecutionContext } from "../../lib/execution-context";
+import { Configuration } from "../../configuration/configuration";
+import { executeChatPipeline } from "../../chat-pipeline/chat-pipeline-completion-api";
+import { executeChatPipeline as executeAssistantPipeline } from "../../chat-pipeline/chat-pipeline-assistant-api";
 
 export async function chat(
   executionContext: ExecutionContext,
