+++ src/commands/commands.ts
@@ -1,0 +1,9 @@
+//  These are the commands that the user can run from the CLI.
+export enum Commands {
+  Unknown,
+  Init,
+  Check,
+  Config,
+  Chat,
+  Quit,
+}
