--- src/fixtures/boxes.json
@@ -1,36 +1,0 @@
-{
-  "boxes": {
-    "steambox": {
-      "commands": {
-        "connect": {
-          "command": "dcv://Administrator@${host}:8443",
-          "copyCommand": "password"
-        }
-      }
-    },
-    "torrentbox": {
-    },
-    "ubox": {
-      "commands": {
-        "connect": {
-          "command": "open vnc://ubuntu@${host}:5901",
-          "copyCommand": "password"
-        },
-        "ssh": {
-          "command": "ssh -i /Users/dwmkerr/repos/github/dwmkerr/dwmkerr/tf-aws-dwmkerr/dwmkerr_aws_key.pem ubuntu@${host}",
-          "copyCommand": "password"
-        }
-      }
-    }
-  },
-  "aws": {
-    "region": "us-west-2"
-  },
-  "commands": {
-    "ssh": {
-      "command": "ssh ${host}",
-      "copyCommand": "ssh ${host}"
-    }
-  }
-}
-
