--- src/lib/constants.ts
@@ -1,4 +1,0 @@
-export const tagNames = {
-  boxId: "boxes.boxid",
-  volumeArchives: "boxes.volumearchives",
-};
